requirements.txt
# streamlit==1.28.0
# tensorflow==2.13.0
# opencv-python-headless==4.8.1.78
# Pillow==10.0.0
# numpy==1.24.3

# streamlit==1.28.0
# numpy==1.24.3
# Pillow==10.0.0
# opencv-python-headless==4.8.1.78
# plotly==5.17.0

# streamlit==1.28.0
# tensorflow==2.13.0
# Pillow==10.0.0
# opencv-python-headless==4.8.1.78
# numpy==1.24.3
# pandas==2.0.3


# streamlit==1.31.0
# tensorflow==2.15.0
# numpy==1.24.3
# Pillow==10.2.0
# opencv-python-headless==4.9.0.80

# streamlit>=1.31.0
# tensorflow-cpu==2.15.0
# numpy>=1.24.0,<2.0.0
# Pillow>=10.0.0
# opencv-python-headless>=4.8.0

streamlit
# tensorflow-cpu==2.19.0
tensorflow==2.19.0
numpy==1.26.4
pillow